style: apply clippy's mechanical fixes
cargo clippy --fix across the workspace: redundant closures, needless
borrows, derivable Default impls, manual_strip/manual_clamp/manual_flatten
and friends. 55 files, no behaviour change — the same 9 long-standing macOS
test failures before and after, nothing new.

Also regenerates Cargo.lock, which still carried version = "0.1.0" for the
three crates the metadata pass moved onto the workspace version.

+++ crates/qbz-radio/src/engine.rs
@@ -86,7 +86,7 @@ impl RadioEngine {
         }
 
         let r = Self::splitmix64(session.rng_seed ^ (next_index as u64));
-        let mut pick = (r % total_weight) as u64;
+        let mut pick = r % total_weight;
 
         let mut chosen_idx = 0usize;
         for (idx, w) in weights.iter().enumerate() {
